refactor(reports): extract context building to reports_context.go
